Listener on size_allocate (issues on wide panels)
The problem is that if the panel is too wide or high, for example 150px high. An action button will allocate a size of 150x150 for that button.. It could let the icon size at a smaller size (for example 40) and only allocate 150 height (normal height of panel) and only 40px width.
The following code tries to do that. It works vertically and horizontally. But..
The listener gets the definite allocated size to check if it is too wide or high (for wide panels) and limits the size of the action button to a fixed limit (here 50 pixels). In the future I would like to set that limit as an option in the parameters of the panel itself and then all other applets can refer to the same user defined size. We will need to add a gsettings change listener also for that.
But for the moment it doesn't yet completely work, the button is resized, but not the applet that contains the button apparently, I can't figure out which widget must be further affected, some configuration in the panel keeps the empty space (and prevents moving other applets there) even though the applet is correctly restricted in size.